fixed image

+++ app/Http/Controllers/Admin/Album/Api/DestroyController.php
@@ -11,13 +11,13 @@ class DestroyController extends BaseController
 {
     public function __invoke(Album $album, User $user)
     {
-        $path = $album->artwork_url;
-        $path = substr($path, 0, strpos($path, basename($path)));
-        $path = pathToServer() . substr($path, strpos($path, "images"));
 
-        $path = preg_replace('/album_artWork\//', '', $path);
-
-        $this->service->delete($path);
+        if (($path = $album->artwork_url) != '') {
+            $path = substr($path, 0, strpos($path, basename($path)));
+            $path = pathToServer() . substr($path, strpos($path, "images"));
+            $path = preg_replace('/album_artWork\//', '', $path);
+            $this->service->delete($path);
+        }
 
         $album->tracks()->detach();
         $album->artists()->detach();
