+++ app/Http/Controllers/Admin/Playlist/Api/DestroyController.php
@@ -11,17 +11,16 @@ class DestroyController extends BaseController
 {
     public function __invoke(Playlist $playlist, User $user)
     {
-        $path = $playlist->thumb_url;
-        $path = substr($path, 0, strpos($path,  basename($path)));
-        $path = pathToServer() . substr($path, strpos($path, "images"));
-
-        $this->service->delete($path);
-
-        $path = $playlist->artwork_url;
-        $path = substr($path, 0, strpos($path,  basename($path)));
-        $path = pathToServer() . substr($path, strpos($path, "images"));
-
-        $this->service->delete($path);
+        if (($path = $playlist->thumb_url)) {
+            $path = substr($path, 0, strpos($path,  basename($path)));
+            $path = pathToServer() . substr($path, strpos($path, "images"));
+            $this->service->delete($path);
+        }
+        if (($path = $playlist->artwork_url) != '') {
+            $path = substr($path, 0, strpos($path,  basename($path)));
+            $path = pathToServer() . substr($path, strpos($path, "images"));
+            $this->service->delete($path);
+        }
 
         $playlist->tracks()->detach();
         $playlist->genres()->detach();
