+++ app/Http/Controllers/Admin/Track/Api/DestroyController.php
@@ -11,13 +11,13 @@ class DestroyController extends BaseController
 {
     public function __invoke(Track $track, User $user)
     {
-        $path = $track->artwork_url;
-        $path = substr($path, 0, strpos($path, basename($path)));
-        $path = pathToServer() . substr($path, strpos($path, "images"));
+        if (($path = $track->artwork_url)) {
+            $path = substr($path, 0, strpos($path, basename($path)));
+            $path = pathToServer() . substr($path, strpos($path, "images"));
+            $path = preg_replace('/artwork\//', '', $path);
+            $this->service->delete($path);
+        }
 
-        $path = preg_replace('/artwork\//', '', $path);
-
-        $this->service->delete($path);
 
         $track->artists()->detach();
         $track->genres()->detach();
